+++ downstairs/src/region.rs
@@ -1208,45 +1208,11 @@ fn config_path<P: AsRef<Path>>(dir: P) -> PathBuf {
     out
 }
 
-/// Verify that the requested block offset and size of the buffer
-/// will fit within the extent.
-pub(crate) fn check_input(
-    extent_size: Block,
-    offset: Block,
-    data: &[u8],
-) -> Result<(), CrucibleError> {
-    let block_size = extent_size.block_size_in_bytes() as u64;
-    /*
-     * Only accept block sized operations
-     */
-    if data.len() != block_size as usize {
-        crucible_bail!(DataLenUnaligned);
-    }
-
-    if offset.block_size_in_bytes() != block_size as u32 {
-        crucible_bail!(BlockSizeMismatch);
-    }
-
-    if offset.shift != extent_size.shift {
-        crucible_bail!(BlockSizeMismatch);
-    }
-
-    let total_size = block_size * extent_size.value;
-    let byte_offset = offset.value * block_size;
-
-    if (byte_offset + data.len() as u64) > total_size {
-        crucible_bail!(OffsetInvalid);
-    }
-
-    Ok(())
-}
-
 #[cfg(test)]
-mod test {
+pub(crate) mod test {
     use std::fs::rename;
     use std::path::PathBuf;
 
-    use bytes::{BufMut, BytesMut};
     use rand::{Rng, RngCore};
     use tempfile::tempdir;
     use uuid::Uuid;
@@ -1254,9 +1220,8 @@ mod test {
     use crate::dump::dump_region;
     use crate::extent::{
         completed_dir, copy_dir, extent_path, remove_copy_cleanup_dir,
-        DownstairsBlockContext, ExtentInner,
+        DownstairsBlockContext,
     };
-    use crucible_protocol::EncryptionContext;
 
     use super::*;
 
@@ -1270,25 +1235,12 @@ mod test {
         TEST_UUID_STR.parse().unwrap()
     }
 
-    fn set_dirty_and_block_contexts(
-        inner: &dyn ExtentInner,
-        block_contexts: &[&DownstairsBlockContext],
-    ) -> Result<()> {
-        inner.set_dirty()?;
-
-        for block_context in block_contexts {
-            inner.set_block_context(block_context)?;
-        }
-
-        Ok(())
-    }
-
     // Create a simple logger
     fn csl() -> Logger {
         build_logger()
     }
 
-    fn new_region_options() -> crucible_common::RegionOptions {
+    pub fn new_region_options() -> crucible_common::RegionOptions {
         let mut region_options: crucible_common::RegionOptions =
             Default::default();
         let block_size = 512;
@@ -1915,168 +1867,10 @@ mod test {
     }
 
     #[test]
-    fn extent_io_valid() {
-        let extent_size = Block::new_512(100);
-        let mut data = BytesMut::with_capacity(512);
-        data.put(&[1; 512][..]);
-
-        check_input(extent_size, Block::new_512(0), &data).unwrap();
-        check_input(extent_size, Block::new_512(99), &data).unwrap();
-    }
-
-    #[test]
-    #[should_panic]
-    fn extent_io_non_aligned_large() {
-        let extent_size = Block::new_512(100);
-        let mut data = BytesMut::with_capacity(513);
-        data.put(&[1; 513][..]);
-
-        check_input(extent_size, Block::new_512(0), &data).unwrap();
-    }
-
-    #[test]
-    #[should_panic]
-    fn extent_io_non_aligned_small() {
-        let extent_size = Block::new_512(100);
-        let mut data = BytesMut::with_capacity(511);
-        data.put(&[1; 511][..]);
-
-        check_input(extent_size, Block::new_512(0), &data).unwrap();
-    }
-
-    #[test]
-    #[should_panic]
-    fn extent_io_bad_block() {
-        let extent_size = Block::new_512(100);
-        let mut data = BytesMut::with_capacity(512);
-        data.put(&[1; 512][..]);
-
-        check_input(extent_size, Block::new_512(100), &data).unwrap();
-    }
-
-    #[test]
-    #[should_panic]
-    fn extent_io_invalid_block_buf() {
-        let extent_size = Block::new_512(100);
-        let mut data = BytesMut::with_capacity(1024);
-        data.put(&[1; 1024][..]);
-
-        check_input(extent_size, Block::new_512(99), &data).unwrap();
-    }
-
-    #[test]
-    #[should_panic]
-    fn extent_io_invalid_large() {
-        let extent_size = Block::new_512(100);
-        let mut data = BytesMut::with_capacity(512 * 100);
-        data.put(&[1; 512 * 100][..]);
-
-        check_input(extent_size, Block::new_512(1), &data).unwrap();
-    }
-
-    #[test]
-    fn extent_name_basic() {
-        assert_eq!(extent_file_name(4, ExtentType::Data), "004");
-    }
-
-    #[test]
-    fn extent_name_basic_ext() {
-        assert_eq!(extent_file_name(4, ExtentType::Db), "004.db");
-    }
-
-    #[test]
-    fn extent_name_basic_ext_shm() {
-        assert_eq!(extent_file_name(4, ExtentType::DbShm), "004.db-shm");
-    }
-
-    #[test]
-    fn extent_name_basic_ext_wal() {
-        assert_eq!(extent_file_name(4, ExtentType::DbWal), "004.db-wal");
-    }
-
-    #[test]
-    fn extent_name_basic_two() {
-        assert_eq!(extent_file_name(10, ExtentType::Data), "00A");
-    }
-
-    #[test]
-    fn extent_name_basic_three() {
-        assert_eq!(extent_file_name(59, ExtentType::Data), "03B");
-    }
-
-    #[test]
-    fn extent_name_max() {
-        assert_eq!(extent_file_name(u32::MAX, ExtentType::Data), "FFF");
-    }
-
-    #[test]
-    fn extent_name_min() {
-        assert_eq!(extent_file_name(u32::MIN, ExtentType::Data), "000");
-    }
-
-    #[test]
-    fn extent_dir_basic() {
-        assert_eq!(extent_dir("/var/region", 4), p("/var/region/00/000/"));
-    }
-
-    #[test]
-    fn extent_dir_max() {
-        assert_eq!(
-            extent_dir("/var/region", u32::MAX),
-            p("/var/region/FF/FFF")
-        );
-    }
-
-    #[test]
-    fn extent_dir_min() {
-        assert_eq!(
-            extent_dir("/var/region", u32::MIN),
-            p("/var/region/00/000/")
-        );
-    }
-
-    #[test]
-    fn extent_path_min() {
-        assert_eq!(
-            extent_path("/var/region", u32::MIN),
-            p("/var/region/00/000/000")
-        );
-    }
-
-    #[test]
     fn copy_path_basic() {
         assert_eq!(
             copy_dir("/var/region", 4),
             p("/var/region/00/000/004.copy")
-        );
-    }
-
-    #[test]
-    fn extent_path_three() {
-        assert_eq!(extent_path("/var/region", 3), p("/var/region/00/000/003"));
-    }
-
-    #[test]
-    fn extent_path_mid_hi() {
-        assert_eq!(
-            extent_path("/var/region", 65536),
-            p("/var/region/00/010/000")
-        );
-    }
-
-    #[test]
-    fn extent_path_mid_lo() {
-        assert_eq!(
-            extent_path("/var/region", 65535),
-            p("/var/region/00/00F/FFF")
-        );
-    }
-
-    #[test]
-    fn extent_path_max() {
-        assert_eq!(
-            extent_path("/var/region", u32::MAX),
-            p("/var/region/FF/FFF/FFF")
         );
     }
 
@@ -2173,415 +1967,6 @@ mod test {
          * Dump the region
          */
         dump_region(dvec, Some(2), None, false, false, csl()).await?;
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn encryption_context() -> Result<()> {
-        let dir = tempdir()?;
-        let mut region =
-            Region::create(&dir, new_region_options(), csl()).await?;
-        region.extend(1).await?;
-
-        let ext = region.get_opened_extent(0).await;
-        let inner = ext.lock().await;
-
-        // Encryption context for blocks 0 and 1 should start blank
-
-        assert!(inner.get_block_contexts(0, 1)?[0].is_empty());
-        assert!(inner.get_block_contexts(1, 1)?[0].is_empty());
-
-        // Set and verify block 0's context
-        set_dirty_and_block_contexts(
-            inner.as_ref(),
-            &[&DownstairsBlockContext {
-                block_context: BlockContext {
-                    encryption_context: Some(EncryptionContext {
-                        nonce: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
-                        tag: [
-                            4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
-                            18, 19,
-                        ],
-                    }),
-                    hash: 123,
-                },
-                block: 0,
-                on_disk_hash: 456,
-            }],
-        )?;
-
-        let ctxs = inner.get_block_contexts(0, 1)?[0].clone();
-
-        assert_eq!(ctxs.len(), 1);
-
-        assert_eq!(
-            ctxs[0]
-                .block_context
-                .encryption_context
-                .as_ref()
-                .unwrap()
-                .nonce,
-            [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]
-        );
-        assert_eq!(
-            ctxs[0]
-                .block_context
-                .encryption_context
-                .as_ref()
-                .unwrap()
-                .tag,
-            [4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19]
-        );
-        assert_eq!(ctxs[0].block_context.hash, 123);
-        assert_eq!(ctxs[0].on_disk_hash, 456);
-
-        // Block 1 should still be blank
-
-        assert!(inner.get_block_contexts(1, 1)?[0].is_empty());
-
-        // Set and verify a new context for block 0
-
-        let blob1 = rand::thread_rng().gen::<[u8; 12]>();
-        let blob2 = rand::thread_rng().gen::<[u8; 16]>();
-
-        // Set and verify block 0's context
-        set_dirty_and_block_contexts(
-            inner.as_ref(),
-            &[&DownstairsBlockContext {
-                block_context: BlockContext {
-                    encryption_context: Some(EncryptionContext {
-                        nonce: blob1,
-                        tag: blob2,
-                    }),
-                    hash: 1024,
-                },
-                block: 0,
-                on_disk_hash: 65536,
-            }],
-        )?;
-
-        let ctxs = inner.get_block_contexts(0, 1)?[0].clone();
-
-        assert_eq!(ctxs.len(), 2);
-
-        // First context didn't change
-        assert_eq!(
-            ctxs[0]
-                .block_context
-                .encryption_context
-                .as_ref()
-                .unwrap()
-                .nonce,
-            [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]
-        );
-        assert_eq!(
-            ctxs[0]
-                .block_context
-                .encryption_context
-                .as_ref()
-                .unwrap()
-                .tag,
-            [4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19]
-        );
-        assert_eq!(ctxs[0].block_context.hash, 123);
-        assert_eq!(ctxs[0].on_disk_hash, 456);
-
-        // Second context was appended
-        assert_eq!(
-            ctxs[1]
-                .block_context
-                .encryption_context
-                .as_ref()
-                .unwrap()
-                .nonce,
-            blob1
-        );
-        assert_eq!(
-            ctxs[1]
-                .block_context
-                .encryption_context
-                .as_ref()
-                .unwrap()
-                .tag,
-            blob2
-        );
-        assert_eq!(ctxs[1].block_context.hash, 1024);
-        assert_eq!(ctxs[1].on_disk_hash, 65536);
-
-        // "Flush", so only the rows that match should remain.
-        inner.truncate_encryption_contexts_and_hashes(&[(0, 65536)])?;
-
-        let ctxs = inner.get_block_contexts(0, 1)?[0].clone();
-
-        assert_eq!(ctxs.len(), 1);
-
-        assert_eq!(
-            ctxs[0]
-                .block_context
-                .encryption_context
-                .as_ref()
-                .unwrap()
-                .nonce,
-            blob1
-        );
-        assert_eq!(
-            ctxs[0]
-                .block_context
-                .encryption_context
-                .as_ref()
-                .unwrap()
-                .tag,
-            blob2
-        );
-        assert_eq!(ctxs[0].block_context.hash, 1024);
-        assert_eq!(ctxs[0].on_disk_hash, 65536);
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn duplicate_context_insert() -> Result<()> {
-        let dir = tempdir()?;
-        let mut region =
-            Region::create(&dir, new_region_options(), csl()).await?;
-        region.extend(1).await?;
-
-        let ext = region.get_opened_extent(0).await;
-        let inner = ext.lock().await;
-
-        assert!(inner.get_block_contexts(0, 1)?[0].is_empty());
-
-        // Submit the same contents for block 0 over and over, simulating a user
-        // writing the same unencrypted contents to the same offset.
-
-        for _ in 0..10 {
-            set_dirty_and_block_contexts(
-                inner.as_ref(),
-                &[&DownstairsBlockContext {
-                    block_context: BlockContext {
-                        encryption_context: None,
-                        hash: 123,
-                    },
-                    block: 0,
-                    on_disk_hash: 123,
-                }],
-            )?;
-        }
-
-        // Duplicate rows should not be inserted
-
-        let ctxs = inner.get_block_contexts(0, 1)?[0].clone();
-        assert_eq!(ctxs.len(), 1);
-
-        // Truncation should still work
-
-        inner.truncate_encryption_contexts_and_hashes(&[(0, 123)])?;
-
-        let ctxs = inner.get_block_contexts(0, 1)?[0].clone();
-        assert_eq!(ctxs.len(), 1);
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn multiple_context() -> Result<()> {
-        let dir = tempdir()?;
-        let mut region =
-            Region::create(&dir, new_region_options(), csl()).await?;
-        region.extend(1).await?;
-
-        let ext = region.get_opened_extent(0).await;
-        let inner = ext.lock().await;
-
-        // Encryption context for blocks 0 and 1 should start blank
-
-        assert!(inner.get_block_contexts(0, 1)?[0].is_empty());
-        assert!(inner.get_block_contexts(1, 1)?[0].is_empty());
-
-        // Set block 0's and 1's context
-        set_dirty_and_block_contexts(
-            inner.as_ref(),
-            &[
-                &DownstairsBlockContext {
-                    block_context: BlockContext {
-                        encryption_context: Some(EncryptionContext {
-                            nonce: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
-                            tag: [
-                                4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16,
-                                17, 18, 19,
-                            ],
-                        }),
-                        hash: 123,
-                    },
-                    block: 0,
-                    on_disk_hash: 456,
-                },
-                &DownstairsBlockContext {
-                    block_context: BlockContext {
-                        encryption_context: Some(EncryptionContext {
-                            nonce: [4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15],
-                            tag: [
-                                8, 9, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12,
-                                13,
-                            ],
-                        }),
-                        hash: 9999,
-                    },
-                    block: 1,
-                    on_disk_hash: 1234567890,
-                },
-            ],
-        )?;
-
-        // Verify block 0's context
-
-        let ctxs = inner.get_block_contexts(0, 1)?[0].clone();
-
-        assert_eq!(ctxs.len(), 1);
-
-        assert_eq!(
-            ctxs[0]
-                .block_context
-                .encryption_context
-                .as_ref()
-                .unwrap()
-                .nonce,
-            [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]
-        );
-        assert_eq!(
-            ctxs[0]
-                .block_context
-                .encryption_context
-                .as_ref()
-                .unwrap()
-                .tag,
-            [4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19]
-        );
-        assert_eq!(ctxs[0].block_context.hash, 123);
-        assert_eq!(ctxs[0].on_disk_hash, 456);
-
-        // Verify block 1's context
-
-        let ctxs = inner.get_block_contexts(1, 1)?[0].clone();
-
-        assert_eq!(ctxs.len(), 1);
-
-        assert_eq!(
-            ctxs[0]
-                .block_context
-                .encryption_context
-                .as_ref()
-                .unwrap()
-                .nonce,
-            [4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15]
-        );
-        assert_eq!(
-            ctxs[0]
-                .block_context
-                .encryption_context
-                .as_ref()
-                .unwrap()
-                .tag,
-            [8, 9, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13]
-        );
-        assert_eq!(ctxs[0].block_context.hash, 9999);
-        assert_eq!(ctxs[0].on_disk_hash, 1234567890);
-
-        // Return both block 0's and block 1's context, and verify
-
-        let ctxs = inner.get_block_contexts(0, 2)?;
-
-        assert_eq!(ctxs[0].len(), 1);
-        assert_eq!(
-            ctxs[0][0]
-                .block_context
-                .encryption_context
-                .as_ref()
-                .unwrap()
-                .nonce,
-            [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]
-        );
-        assert_eq!(
-            ctxs[0][0]
-                .block_context
-                .encryption_context
-                .as_ref()
-                .unwrap()
-                .tag,
-            [4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19]
-        );
-        assert_eq!(ctxs[0][0].block_context.hash, 123);
-        assert_eq!(ctxs[0][0].on_disk_hash, 456);
-
-        assert_eq!(ctxs[1].len(), 1);
-        assert_eq!(
-            ctxs[1][0]
-                .block_context
-                .encryption_context
-                .as_ref()
-                .unwrap()
-                .nonce,
-            [4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15]
-        );
-        assert_eq!(
-            ctxs[1][0]
-                .block_context
-                .encryption_context
-                .as_ref()
-                .unwrap()
-                .tag,
-            [8, 9, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13]
-        );
-        assert_eq!(ctxs[1][0].block_context.hash, 9999);
-        assert_eq!(ctxs[1][0].on_disk_hash, 1234567890);
-
-        // Append a whole bunch of block context rows
-        for i in 0..10 {
-            set_dirty_and_block_contexts(
-                inner.as_ref(),
-                &[
-                    &DownstairsBlockContext {
-                        block_context: BlockContext {
-                            encryption_context: Some(EncryptionContext {
-                                nonce: rand::thread_rng().gen::<[u8; 12]>(),
-                                tag: rand::thread_rng().gen::<[u8; 16]>(),
-                            }),
-                            hash: rand::thread_rng().gen::<u64>(),
-                        },
-                        block: 0,
-                        on_disk_hash: i,
-                    },
-                    &DownstairsBlockContext {
-                        block_context: BlockContext {
-                            encryption_context: Some(EncryptionContext {
-                                nonce: rand::thread_rng().gen::<[u8; 12]>(),
-                                tag: rand::thread_rng().gen::<[u8; 16]>(),
-                            }),
-                            hash: rand::thread_rng().gen::<u64>(),
-                        },
-                        block: 1,
-                        on_disk_hash: i,
-                    },
-                ],
-            )?;
-        }
-
-        let ctxs = inner.get_block_contexts(0, 2)?;
-        assert_eq!(ctxs[0].len(), 11);
-        assert_eq!(ctxs[1].len(), 11);
-
-        // "Flush", so only the rows that match the on-disk hash should remain.
-
-        inner.truncate_encryption_contexts_and_hashes(&[(0, 6), (1, 7)])?;
-
-        let ctxs = inner.get_block_contexts(0, 2)?;
-
-        assert_eq!(ctxs[0].len(), 1);
-        assert_eq!(ctxs[0][0].on_disk_hash, 6);
-
-        assert_eq!(ctxs[1].len(), 1);
-        assert_eq!(ctxs[1][0].on_disk_hash, 7);
 
         Ok(())
     }
@@ -2689,21 +2074,19 @@ mod test {
             Region::create(&dir, new_region_options(), csl()).await?;
         region.extend(1).await?;
 
-        // A write of some sort only wrote a block context row
+        // A write of some sort only wrote a block context row and dirty flag
         {
             let ext = region.get_opened_extent(0).await;
             let inner = ext.lock().await;
-            set_dirty_and_block_contexts(
-                inner.as_ref(),
-                &[&DownstairsBlockContext {
-                    block_context: BlockContext {
-                        encryption_context: None,
-                        hash: 1024,
-                    },
-                    block: 0,
-                    on_disk_hash: 65536,
-                }],
-            )?;
+            inner.set_dirty()?;
+            inner.set_block_context(&DownstairsBlockContext {
+                block_context: BlockContext {
+                    encryption_context: None,
+                    hash: 1024,
+                },
+                block: 0,
+                on_disk_hash: 65536,
+            })?;
         }
 
         // This should clear out the invalid contexts
@@ -2754,188 +2137,6 @@ mod test {
         let response = &responses[0];
 
         assert_eq!(response.data.to_vec(), vec![0x55; 512]);
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    /// It's very important that only_write_unwritten always works correctly.
-    /// We should never add contexts for blocks that haven't been written by
-    /// an upstairs.
-    async fn test_fully_rehash_and_clean_does_not_mark_blocks_as_written(
-    ) -> Result<()> {
-        // Make a region. It's empty, and should stay that way
-        let dir = tempdir()?;
-        let mut region =
-            Region::create(&dir, new_region_options(), csl()).await?;
-        region.extend(1).await?;
-
-        let ext = region.get_opened_extent(0).await;
-
-        // Write a block, but don't flush.
-        {
-            let data = Bytes::from(vec![0x55; 512]);
-            let hash = integrity_hash(&[&data[..]]);
-            let write = crucible_protocol::Write {
-                eid: 0,
-                offset: Block::new_512(0),
-                data,
-                block_context: BlockContext {
-                    encryption_context: None,
-                    hash,
-                },
-            };
-            ext.write(JobId(10), &[&write], false).await?;
-        }
-
-        // We haven't flushed, but this should leave our context in place.
-        ext.fully_rehash_and_clean_all_stale_contexts(false).await?;
-
-        // Therefore, we expect that write_unwritten to the first block won't
-        // do anything.
-        {
-            let data = Bytes::from(vec![0x66; 512]);
-            let hash = integrity_hash(&[&data[..]]);
-            let block_context = BlockContext {
-                encryption_context: None,
-                hash,
-            };
-            let write = crucible_protocol::Write {
-                eid: 0,
-                offset: Block::new_512(0),
-                data: data.clone(),
-                block_context: block_context.clone(),
-            };
-            ext.write(JobId(20), &[&write], true).await?;
-
-            let mut resp = Vec::new();
-            let read = ReadRequest {
-                eid: 0,
-                offset: Block::new_512(0),
-            };
-            ext.read(JobId(21), &[&read], &mut resp).await?;
-
-            // We should not get back our data, because block 0 was written.
-            assert_ne!(
-                resp,
-                vec![ReadResponse {
-                    eid: 0,
-                    offset: Block::new_512(0),
-                    data: BytesMut::from(data.as_ref()),
-                    block_contexts: vec![block_context]
-                }]
-            );
-        }
-
-        // But, writing to the second block still should!
-        {
-            let data = Bytes::from(vec![0x66; 512]);
-            let hash = integrity_hash(&[&data[..]]);
-            let block_context = BlockContext {
-                encryption_context: None,
-                hash,
-            };
-            let write = crucible_protocol::Write {
-                eid: 0,
-                offset: Block::new_512(1),
-                data: data.clone(),
-                block_context: block_context.clone(),
-            };
-            ext.write(JobId(30), &[&write], true).await?;
-
-            let mut resp = Vec::new();
-            let read = ReadRequest {
-                eid: 0,
-                offset: Block::new_512(1),
-            };
-            ext.read(JobId(31), &[&read], &mut resp).await?;
-
-            // We should get back our data! Block 1 was never written.
-            assert_eq!(
-                resp,
-                vec![ReadResponse {
-                    eid: 0,
-                    offset: Block::new_512(1),
-                    data: BytesMut::from(data.as_ref()),
-                    block_contexts: vec![block_context]
-                }]
-            );
-        }
-
-        Ok(())
-    }
-
-    /// If a write successfully put a context into the database, but it never
-    /// actually got the data onto the disk, that block should revert back to
-    /// being "unwritten". After all, the data never was truly written.
-    ///
-    /// This test is very similar to test_region_open_removes_partial_writes,
-    /// but is distinct in that it call fully_rehash directly, without closing
-    /// and re-opening the extent.
-    #[tokio::test]
-    async fn test_fully_rehash_marks_blocks_unwritten_if_data_never_hit_disk(
-    ) -> Result<()> {
-        let dir = tempdir()?;
-        let mut region =
-            Region::create(&dir, new_region_options(), csl()).await?;
-        region.extend(1).await?;
-
-        let ext = region.get_opened_extent(0).await;
-
-        // Partial write, the data never hits disk, but there's a context
-        // in the DB
-        {
-            let inner = ext.lock().await;
-            set_dirty_and_block_contexts(
-                inner.as_ref(),
-                &[&DownstairsBlockContext {
-                    block_context: BlockContext {
-                        encryption_context: None,
-                        hash: 1024,
-                    },
-                    block: 0,
-                    on_disk_hash: 65536,
-                }],
-            )?;
-        }
-
-        // Run a full rehash, which should clear out that partial write.
-        ext.fully_rehash_and_clean_all_stale_contexts(false).await?;
-
-        // Writing to block 0 should succeed with only_write_unwritten
-        {
-            let data = Bytes::from(vec![0x66; 512]);
-            let hash = integrity_hash(&[&data[..]]);
-            let block_context = BlockContext {
-                encryption_context: None,
-                hash,
-            };
-            let write = crucible_protocol::Write {
-                eid: 0,
-                offset: Block::new_512(0),
-                data: data.clone(),
-                block_context: block_context.clone(),
-            };
-            ext.write(JobId(30), &[&write], true).await?;
-
-            let mut resp = Vec::new();
-            let read = ReadRequest {
-                eid: 0,
-                offset: Block::new_512(0),
-            };
-            ext.read(JobId(31), &[&read], &mut resp).await?;
-
-            // We should get back our data! Block 1 was never written.
-            assert_eq!(
-                resp,
-                vec![ReadResponse {
-                    eid: 0,
-                    offset: Block::new_512(0),
-                    data: BytesMut::from(data.as_ref()),
-                    block_contexts: vec![block_context]
-                }]
-            );
-        }
 
         Ok(())
     }
